Added further test cases for Goal, changes in previous test cases, code refactoring
